mpc is cooler now

+++ anergos.d/mpc.sh
@@ -1,44 +1,66 @@
 #!/usr/bin/env bash
 
+function check_val() { 
+	if [ $1 ] && [ $1 -eq 0 ]; then
+		echo $(tput setaf 1)"${FUNCNAME[1]} skipped"$(tput sgr0) && return 69
+	fi
+}
+
 function nobeep() {
+	check_val $1; [ $? -eq 69 ] && return
 	echo "blacklist pcspkr" > /etc/modprobe.d/nobeep.conf
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function all_core_make() {
+	check_val $1; [ $? -eq 69 ] && return
 	grep "^MAKEFLAGS" /etc/makepkg.conf >/dev/null 2>&1 || 
 	sed -i "s/-j2/-j$(nproc)/;s/^#MAKEFLAGS/MAKEFLAGS/" /etc/makepkg.conf
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function nano_configs() {
+	check_val $1; [ $? -eq 69 ] && return
 	grep '^include "/usr/share/nano/*.nanorc"' /etc/nanorc >/dev/null 2>&1 || 
 	echo 'include "/usr/share/nano/*.nanorc"' >> /etc/nanorc
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function infinality(){
+	check_val $1; [ $? -eq 69 ] && return
 	sed -i 's/^#exp/exp/;s/version=40"$/version=38"$/' /etc/profile.d/freetype2.sh
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function office_logo() {
-	[ -f /etc/libreoffice/sofficerc ] && sed -i 's/Logo=1/Logo=0/g' /etc/libreoffice/sofficerc; }
+	check_val $1; [ $? -eq 69 ] && return
+	[ -f /etc/libreoffice/sofficerc ] && sed -i 's/Logo=1/Logo=0/g' /etc/libreoffice/sofficerc
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
+}
 
 function create_swapfile() {
+	check_val $1; [ $? -eq 69 ] && return
 	fallocate -l 2G /swapfile 	>/dev/null 2>&1
 	chmod 600 /swapfile
 	mkswap /swapfile 			>/dev/null 2>&1
 	swapon /swapfile 			>/dev/null 2>&1
 	printf "# Swapfile\\n/swapfile none swap defaults 0 0\\n\\n" >> /etc/fstab
 	printf "vm.swappiness=10\nvm.vfs_cache_pressure=50\n" > /etc/sysctl.d/99-sysctl.conf
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function clone_dotfiles() {
+	check_val $1; [ $? -eq 69 ] && return
 	cd /home/"$name" && echo ".cfg" >> .gitignore && rm .bash_profile .bashrc
 	sudo -u "$name" git clone --bare "$dotfilesrepo" /home/${name}/.cfg > /dev/null 2>&1 
 	sudo -u "$name" git --git-dir=/home/${name}/.cfg/ --work-tree=/home/${name} checkout
 	sudo -u "$name" git --git-dir=/home/${name}/.cfg/ --work-tree=/home/${name} config \
 				--local status.showUntrackedFiles no > /dev/null 2>&1 && rm .gitignore
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function arduino() {
+	check_val $1; [ $? -eq 69 ] && return
 	if [ -f /usr/bin/arduino ]; then
 		echo cdc_acm > /etc/modules-load.d/cdc_acm.conf
 		sudo -u "$name" groups | grep uucp >/dev/null 2>&1 || gpasswd -a $name uucp >/dev/null 2>&1
@@ -46,9 +68,11 @@ function arduino() {
 	else
 		echo "Skipped."
 	fi
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function virtualbox() {
+	check_val $1; [ $? -eq 69 ] && return
 	if [[ $(lspci | grep VirtualBox) ]] && [ -f /usr/bin/pacman ]; then
 		pacman -S --noconfirm virtualbox-guest-modules-arch virtualbox-guest-utils >/dev/null 2>&1
 	elif [ -f /usr/bin/virtualbox ]; then
@@ -56,31 +80,40 @@ function virtualbox() {
 	else
 		echo "Skipped."
 	fi
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function networkd_config() {
-	systemctl enable NetworkManager >/dev/null 2>&1 || {
-	net_lot=$(networkctl --no-legend | grep -P "ether|wlan" | awk '{print $2}')
-	for device in ${net_lot[*]}; do ((i++))
-		cat > /etc/systemd/network/${device}.network <<-EOF
-			[Match]
-			Name=${device}
-			[Network]
-			DHCP=ipv4
-			[DHCP]
-			RouteMetric=$(($i * 10))
-		EOF
-	done
-	systemctl enable systemd-networkd >/dev/null 2>&1
-	systemctl enable systemd-resolved >/dev/null 2>&1 
-	}
+	check_val $1; [ $? -eq 69 ] && return
+	if [ -f  /usr/bin/NetworkManager ]; then
+		systemctl enable NetworkManager >/dev/null 2>&1
+	else
+		net_lot=$(networkctl --no-legend 2>/dev/null | grep -P "ether|wlan" | awk '{print $2}')
+		for device in ${net_lot[*]}; do ((i++))
+			cat > /etc/systemd/network/${device}.network <<-EOF
+				[Match]
+				Name=${device}
+				[Network]
+				DHCP=ipv4
+				[DHCP]
+				RouteMetric=$(($i * 10))
+			EOF
+		done
+		systemctl enable systemd-networkd >/dev/null 2>&1
+		systemctl enable systemd-resolved >/dev/null 2>&1
+	fi 
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function power_group() {
+	check_val $1; [ $? -eq 69 ] && return
 	gpasswd -a $name power >/dev/null 2>&1
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function agetty_set() {
+	systemctl enable gdm >/dev/null 2>&1 && echo "Enabled gdm" && return
+	check_val $1; [ $? -eq 69 ] && return
 	if [ "$1" = "auto" ]; then
 		log="ExecStart=-\/sbin\/agetty --autologin $name --noclear %I \$TERM"
 	else
@@ -89,9 +122,12 @@ function agetty_set() {
 	sed "s/ExecStart=.*/${log}/" /usr/lib/systemd/system/getty@.service > /etc/systemd/system/getty@.service
 	systemctl daemon-reload 				>/dev/null 2>&1
 	systemctl reenable getty@tty1.service 	>/dev/null 2>&1
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function lock_sleep() {
+	systemctl enable gdm >/dev/null 2>&1 && echo "Enabled gdm" && return
+	check_val $1; [ $? -eq 69 ] && return
 	if [ -f /usr/bin/i3lock ] && [ ! -f /usr/bin/sway ]; then
 		cat > /etc/systemd/system/SleepLocki3@${name}.service <<-EOF
 			#/etc/systemd/system/
@@ -109,13 +145,17 @@ function lock_sleep() {
 		EOF
 	fi
 	systemctl enable SleepLocki3@${name}
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function resolv_conf() {
+	check_val $1; [ $? -eq 69 ] && return
 	printf "search home\\nnameserver 192.168.1.1\\n" > /etc/resolv.conf
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function enable_numlk_tty() {
+	check_val $1; [ $? -eq 69 ] && return
 	cat > /etc/systemd/system/numLockOnTty.service <<-EOF
 		[Unit]
 		Description=numlockOnTty
@@ -135,49 +175,56 @@ function enable_numlk_tty() {
 	EOF
 	chmod +x /usr/bin/numlockOnTty
 	systemctl enable numLockOnTty >/dev/null 2>&1
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function temps() {
+	check_val $1; [ $? -eq 69 ] && return
 	if [ -f /usr/bin/yay ]; then
 		# https://aur.archlinux.org/packages/it87-dkms-git/ || https://github.com/bbqlinux/it87
 		sudo -u "$name" yay -S --noconfirm it87-dkms-git >/dev/null 2>&1
 		echo "it87" > /etc/modules-load.d/it87.conf
 		echo "it87-dkms-git" >> /home/"name"/.local/Fresh_pack_list
 	fi
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function data() {
+	check_val $1; [ $? -eq 69 ] && return
 	mkdir -p /media/Data
 	cat >> /etc/fstab <<-EOF
 		# /dev/sda1 LABEL=data
 		UUID=fe8b7dcf-3bae-4441-a4f3-a3111fee8ca4  /media/Data    ext4 rw,noatime,nofail,user,auto    0  2
 	
 	EOF
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 function powerb_is_suspend() {
+	check_val $1; [ $? -eq 69 ] && return
 	sed -i "s/^#HandlePowerKey=poweroff/HandlePowerKey=suspend/g" /etc/systemd/logind.conf
+	echo $(tput setaf 2)"${FUNCNAME[0]} done \( value $1\)"$(tput sgr0)
 }
 
 # Needed Permissions
 echo "%wheel ALL=(ALL) NOPASSWD: ALL" > /etc/sudoers.d/wheel && chmod 440 /etc/sudoers.d/wheel
 
-nobeep
-all_core_make
-nano_configs
-infinality
-office_logo
-create_swapfile 				
-clone_dotfiles
-arduino
-virtualbox
-networkd_config
-power_group
+nobeep 				1
+all_core_make 		1
+nano_configs 		1
+infinality 			1
+office_logo 		1
+create_swapfile 	1
+clone_dotfiles 		1
+arduino 			1
+virtualbox 			1
+networkd_config 	1
+power_group 		1
+agetty_set			1
+lock_sleep			1
 
-systemctl enable gdm >/dev/null 2>&1 || agetty_set && lock_sleep
-
-[ $hostname = "killua" ] && { resolv_conf; enable_numlk_tty; temps; data; }
-[ $hostname = "killua" ] && powerb_is_suspend 
+[ $hostname = "killua" ] && { resolv_conf 1; enable_numlk_tty 1; temps 1; data 1; }
+[ $hostname = "killua" ] && powerb_is_suspend 0
 
 # Sane Permissions
 cat > /etc/sudoers.d/wheel <<-EOF
